refactor: move recipient extraction logic to a new utility module and update RecipientAnalysisService to use it
